+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Unreleased
 * Feature: Selection of predefined time periods
+* Enhancement: Better selection of posts
 * Enhancement: Minified CSS/JS files, build command
 * Enhancement: Check for compliance with WordPress Coding Guidelines with PHP_CodeSniffer
 * Refactoring the charts code